backend/requirements.txt
# ── Shared Infrastructure ─────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0

# ── Query System (System 1) ───────────────────────────────────────────────────
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0   # local embeddings — no data sent externally
chromadb>=0.5.0
pymupdf>=1.24.0                        # fast PDF text extraction (parser.py first-choice)
python-docx>=1.1.0                     # DOCX text extraction (parser.py second-choice)
unstructured[pdf,docx]>=0.13.0        # fallback parser for unsupported formats

# ── Feedback System (System 2) ────────────────────────────────────────────────
crewai>=0.28.0
ollama>=0.2.0
# smtplib ships with Python stdlib — no pip package needed
